+++ tests/test_module_split_compat.py
@@ -171,8 +171,96 @@ SWEEP_NAMES = (
     "write_leaf_submap",
 )
 
+#: Top-level names ``src/zagg/config.py`` defined before the phase-4 split (the
+#: module was 2,747 lines — the second-worst offender).
+CONFIG_NAMES = (
+    "DataSourceDict",
+    "FILTER_OPS",
+    "LevelDict",
+    "LinkDict",
+    "OUTPUT_KINDS",
+    "OUTPUT_RESOLUTIONS",
+    "PipelineConfig",
+    "WORKER_MEMORIES",
+    "_NAN_AMBIGUOUS_REDUCTIONS",
+    "_PIPELINE_TYPES",
+    "_RESAMPLE_HOWS",
+    "_SCALAR_OPS",
+    "_SET_OPS",
+    "_TEMPORAL_SPEC_KEYS",
+    "_eval_expression_raw",
+    "_is_nan_fill",
+    "_is_numeric",
+    "_normalize_filter",
+    "_segment_variable_names",
+    "_validate_chunk_precompute",
+    "_validate_collection_options",
+    "_validate_composition_attrs",
+    "_validate_ds_variables",
+    "_validate_expression_columns",
+    "_validate_filter_levels",
+    "_validate_filters",
+    "_validate_index",
+    "_validate_levels",
+    "_validate_output_kind",
+    "_validate_pyramid",
+    "_validate_raster_config",
+    "_validate_store_layout_keys",
+    "_validate_temporal_config",
+    "_validate_trailing_shape",
+    "_validate_windowing",
+    "_validate_windowing_windows",
+    "_validate_worker",
+    "_variable_entry",
+    "_warn_nan_ambiguous_reductions",
+    "collection_options",
+    "default_config",
+    "evaluate_expression",
+    "evaluate_filter_expression",
+    "filters_from_data_source",
+    "get_agg_fields",
+    "get_aoi_mask",
+    "get_base_level",
+    "get_child_order",
+    "get_chunk_precompute",
+    "get_consolidate_metadata",
+    "get_coords",
+    "get_coverage_moc",
+    "get_data_vars",
+    "get_driver",
+    "get_emit_cell_ids",
+    "get_filters",
+    "get_handoff",
+    "get_layout",
+    "get_levels",
+    "get_output_endpoint_url",
+    "get_output_region",
+    "get_output_signature",
+    "get_parent_order",
+    "get_pipeline_type",
+    "get_product_name",
+    "get_pyramid",
+    "get_raster_bands",
+    "get_sharded",
+    "get_store_layout",
+    "get_store_path",
+    "get_sweep",
+    "get_windowing",
+    "load_config",
+    "load_config_from_dict",
+    "output_field_signature",
+    "resolve_function",
+    "validate_config",
+    "window_time_filters",
+    "windowed_cell_config",
+)
+
 #: ``{facade module: (pre-split names, submodules it re-exports from)}``.
 SPLITS = {
+    "zagg.config": (
+        CONFIG_NAMES,
+        ("base", "expressions", "accessors", "validate_source", "validate_output", "validate"),
+    ),
     "zagg.hive": (HIVE_NAMES, ("layout", "manifest", "coverage")),
     "zagg.processing.raster": (RASTER_NAMES, ("decode", "template", "write")),
     "zagg.sweep": (SWEEP_NAMES, ("zagg.sweep_families", "zagg.sweep_rollup")),
